Simplify class system
