--- src/main java/ru/mail/track/kolodzey/app/ui/MainUI.java
@@ -1,8 +1,0 @@
-package ru.mail.track.kolodzey.app.ui;
-
-/**
- * Created by DKolodzey on 06.10.15.
- */
-public interface MainUI {
-    public void run();
-}
